# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Winzy API
  version: 1.0.0
  description: >
    Hand-written contract for the single Go API service. Source of truth for
    wire shapes is backend/internal/*/ models and handlers — not the retired
    C# microservices. Drift is enforced by backend/cmd/api contract tests.
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /habits:
    get:
      tags: [habits]
      operationId: listHabits
      responses:
        "200":
          description: Active habits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Habit"
    post:
      tags: [habits]
      operationId: createHabit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHabitRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
        "400":
          $ref: "#/components/responses/Error"
  /habits/completions:
    get:
      tags: [habits]
      operationId: completionsByDate
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Completions for date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompletionsByDateResponse"
        "400":
          $ref: "#/components/responses/Error"
  /habits/{id}:
    parameters:
      - $ref: "#/components/parameters/HabitId"
    get:
      tags: [habits]
      operationId: getHabit
      responses:
        "200":
          description: Habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
        "404":
          description: Not found
    put:
      tags: [habits]
      operationId: updateHabit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHabitRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          description: Not found
    delete:
      tags: [habits]
      operationId: archiveHabit
      responses:
        "204":
          description: Archived
        "404":
          description: Not found
  /habits/{id}/complete:
    post:
      tags: [habits]
      operationId: completeHabit
      parameters:
        - $ref: "#/components/parameters/HabitId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteHabitRequest"
      responses:
        "200":
          description: Completion logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HabitCompletion"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          description: Not found
        "409":
          $ref: "#/components/responses/Error"
  /habits/{id}/completions/{date}:
    parameters:
      - $ref: "#/components/parameters/HabitId"
      - name: date
        in: path
        required: true
        schema:
          type: string
          format: date
    put:
      tags: [habits]
      operationId: updateCompletion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCompletionRequest"
      responses:
        "200":
          description: Updated completion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateCompletionResponse"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          description: Not found
    delete:
      tags: [habits]
      operationId: deleteCompletion
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /habits/{id}/promise:
    parameters:
      - $ref: "#/components/parameters/HabitId"
    get:
      tags: [habits]
      operationId: getPromise
      parameters:
        - $ref: "#/components/parameters/XTimezone"
        - name: history
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Active promise (+ optional history)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromiseResponse"
        "404":
          description: Not found
    post:
      tags: [habits]
      operationId: createPromise
      parameters:
        - $ref: "#/components/parameters/XTimezone"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePromiseRequest"
      responses:
        "201":
          description: Created promise
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlamePromise"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          description: Not found
        "409":
          $ref: "#/components/responses/Error"
    delete:
      tags: [habits]
      operationId: cancelPromise
      responses:
        "204":
          description: Cancelled
        "404":
          description: Not found
  /habits/{id}/promise/visibility:
    patch:
      tags: [habits]
      operationId: togglePromiseVisibility
      parameters:
        - $ref: "#/components/parameters/HabitId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TogglePromiseVisibilityRequest"
      responses:
        "200":
          description: Updated visibility
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TogglePromiseVisibilityResponse"
        "404":
          description: Not found
  /habits/public/{username}:
    get:
      tags: [habits]
      operationId: publicFlameProfile
      security: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public flame profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicFlameProfile"
        "404":
          description: Not found
  /habits/public/{username}/flame.svg:
    get:
      tags: [habits]
      operationId: flameBadge
      security: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: SVG flame badge
          content:
            image/svg+xml:
              schema:
                type: string
        "404":
          description: Not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    HabitId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    XTimezone:
      name: X-Timezone
      in: header
      required: true
      schema:
        type: string
      description: IANA timezone id required by owner-facing habit stats/promise endpoints
  responses:
    Error:
      description: Plain error body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
...
  schemas:
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    FrequencyType:
      type: string
      enum: [daily, weekly, custom]
    CompletionKind:
      type: string
      enum: [full, minimum]
    FlameLevel:
      type: string
      enum: [none, ember, steady, strong, blazing]
    Habit:
      type: object
      required:
        [id, name, icon, color, frequency, customDays, minimumDescription, createdAt, archivedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        icon:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        frequency:
          $ref: "#/components/schemas/FrequencyType"
        customDays:
          type: ["array", "null"]
          items:
            type: integer
        minimumDescription:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        archivedAt:
          type: ["string", "null"]
          format: date-time
    CreateHabitRequest:
      type: object
      required: [name, frequency]
      properties:
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        frequency:
          $ref: "#/components/schemas/FrequencyType"
        customDays:
          type: array
          items:
            type: integer
        minimumDescription:
          type: string
    UpdateHabitRequest:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        frequency:
          $ref: "#/components/schemas/FrequencyType"
        customDays:
          type: array
          items:
            type: integer
        minimumDescription:
          type: string
        clearMinimumDescription:
          type: boolean
    CompleteHabitRequest:
      type: object
      required: [timezone]
      properties:
        date:
          type: string
          format: date
        timezone:
          type: string
        completionKind:
          type: integer
          description: "1=full, 2=minimum (numeric wire enum on requests)"
    UpdateCompletionRequest:
      type: object
      required: [completionKind]
      properties:
        completionKind:
          type: integer
          description: "1=full, 2=minimum"
    HabitCompletion:
      type: object
      required: [id, habitId, localDate, completedAt, completionKind, consistency]
      properties:
        id:
          type: string
          format: uuid
        habitId:
          type: string
          format: uuid
        localDate:
          type: string
          format: date
        completedAt:
          type: string
          format: date-time
        completionKind:
          $ref: "#/components/schemas/CompletionKind"
        consistency:
          type: number
    UpdateCompletionResponse:
      type: object
      required: [id, habitId, localDate, completedAt, completionKind]
      properties:
        id:
          type: string
          format: uuid
        habitId:
          type: string
          format: uuid
        localDate:
          type: string
          format: date
        completedAt:
          type: string
          format: date-time
        completionKind:
          $ref: "#/components/schemas/CompletionKind"
...
    CompletionsByDateResponse:
      type: object
      required: [date, habits]
      properties:
        date:
          type: string
          format: date
        habits:
          type: array
          items:
            type: object
            required: [id, name, icon, color, minimumDescription, completed, completionKind]
            properties:
              id:
                type: string
                format: uuid
              name:
                type: string
              icon:
                type: ["string", "null"]
              color:
                type: ["string", "null"]
              minimumDescription:
                type: ["string", "null"]
              completed:
                type: boolean
              completionKind:
                type: ["string", "null"]
    PromiseStatus:
      type: string
      enum: [active, kept, endedbelow, cancelled]
    FlamePromise:
      type: object
      required:
        [
          id,
          habitId,
          targetConsistency,
          endDate,
          privateNote,
          status,
          onTrack,
          currentConsistency,
          isPublicOnFlame,
          statement,
          createdAt,
          resolvedAt,
        ]
      properties:
        id:
          type: string
          format: uuid
        habitId:
          type: string
          format: uuid
        targetConsistency:
          type: number
        endDate:
          type: string
          format: date
        privateNote:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/PromiseStatus"
        onTrack:
          type: ["boolean", "null"]
        currentConsistency:
          type: ["number", "null"]
        isPublicOnFlame:
          type: boolean
        statement:
          type: string
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: ["string", "null"]
          format: date-time
    PublicPromise:
      type: object
      required: [targetConsistency, endDate, statement, onTrack]
      properties:
        targetConsistency:
          type: number
        endDate:
          type: string
          format: date
        statement:
          type: string
        onTrack:
          type: ["boolean", "null"]
    PromiseResponse:
      type: object
      required: [active, history]
      properties:
        active:
          oneOf:
            - $ref: "#/components/schemas/FlamePromise"
            - type: "null"
        history:
          type: array
          items:
            $ref: "#/components/schemas/FlamePromise"
    CreatePromiseRequest:
      type: object
      required: [targetConsistency, endDate]
      properties:
        targetConsistency:
          type: number
        endDate:
          type: string
          format: date
        privateNote:
          type: string
        isPublicOnFlame:
          type: boolean
    TogglePromiseVisibilityRequest:
      type: object
      required: [isPublicOnFlame]
      properties:
        isPublicOnFlame:
          type: boolean
    TogglePromiseVisibilityResponse:
      type: object
      required: [isPublicOnFlame]
      properties:
        isPublicOnFlame:
          type: boolean
    PublicFlameProfile:
      type: object
      required: [username, habits, degraded]
      properties:
        username:
          type: string
        habits:
          type: array
          items:
            type: object
            required: [id, name, icon, color, consistency, flameLevel, promise]
            properties:
              id:
                type: string
                format: uuid
              name:
                type: string
              icon:
                type: ["string", "null"]
              color:
                type: ["string", "null"]
              consistency:
                type: number
              flameLevel:
                $ref: "#/components/schemas/FlameLevel"
              promise:
                oneOf:
                  - $ref: "#/components/schemas/PublicPromise"
                  - type: "null"
        degraded:
          type: boolean
...